sync

+++ test/cstdlib.h
@@ -1,4 +1,4 @@
-*void malloc(int size);
+*void malloc(uint64 size);
 int puts(*char string);
 *void memcpy(*void dest, *void src, uint64 n);
 *char strcpy(*char dest, *char src);
